ml_service/requirements.txt
fastapi==0.115.2
uvicorn==0.30.6
tensorflow==2.20.0
numpy>=1.26.0
Pillow==10.4.0
opencv-python-headless==4.10.0.84
python-multipart==0.0.9